+++ packages/lumx-vue/src/components/flex-box/index.ts
@@ -1,0 +1,14 @@
+import { CLASSNAME } from '@lumx/core/js/components/FlexBox';
+
+import FlexBox from './FlexBox.vue';
+
+FlexBox.className = CLASSNAME;
+
+export { FlexBox };
+export type { FlexBoxProps } from './types';
+export type {
+    MarginAutoAlignment,
+    GapSize,
+    FlexVerticalAlignment,
+    FlexHorizontalAlignment,
+} from '@lumx/core/js/components/FlexBox';